# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/client": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Reads all clients",
                "description": "Returns all clients contained in the database",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Add new client",
                "description": "Add a new customer to the database.",
                "operationId": "create",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Client email",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Client name",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Wrong value"
                    }
                }
            }
        },
        "/client/{id}": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Read client",
                "description": "Returns information about the client identified by 'id'.",
                "operationId": "read",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client identifier",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Client"
                ],
                "summary": "Delete client",
                "description": "Removes the client identified by 'id'.",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client identifier",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/client/{id}/note": {
            "post": {
                "tags": [
                    "Note"
                ],
                "summary": "Add new Note",
                "description": "Add a new note to the database. The 'id' variable represents the identifier of the client concerned.",
                "operationId": "create",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client identifier",
                        "required": true
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Note value",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Wrong value"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/note/{id}": {
            "get": {
                "tags": [
                    "Note"
                ],
                "summary": "Read Note",
                "description": "Returns information about the note identified by 'id'.",
                "operationId": "read",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note identifier",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Note"
                ],
                "summary": "Update Note",
                "description": "Replaces the 'rating' value of the note obtained with the 'id' identifier by the 'rating' parameter.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Note value",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note identifier",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Wrong value"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Note"
                ],
                "summary": "Delete Note",
                "description": "Removes the note identified by 'id'.",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note identifier",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/note": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Reads all notes",
                "description": "Returns all notes contained in the database",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}